chore: Move `Track` back into the `components` folder
- We'll make a new `multiSelect` folder instead.
