+++ mobile/src/navigation/screens/playlists/CurrentView.tsx
@@ -22,7 +22,7 @@ import { IconButton } from "~/components/Form/Button/Icon";
 import { SafeContainer } from "~/components/SafeContainer";
 import { useSheetRef } from "~/components/Sheet/useSheetRef";
 import { FavoritesPlaylistKey } from "~/modules/media/constants";
-import { useTrackListPreset } from "~/modules/media/track/components/Track";
+import { useTrackListPreset } from "~/modules/media/components/Track";
 
 type Props = StaticScreenProps<{ id: string }>;
